requirements.txt
python-dotenv>=0.15.0
sshtunnel>=0.3.1
mysql-connector-python>=8.0.28
